+++ CHANGELOG.md
@@ -1,30 +1,41 @@
 # Change Log
+
 All notable changes to this project will be documented in this file. (http://keepachangelog.com/)
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## unreleased
+
 - Installed Sentry
+- Upgrade Nuxt.js to v.2.13
 
 ## [1.0.6] - 2010-06-02
+
 - Fixed error page
 
 ## [1.0.6] - 2010-06-02
+
 - Added missing translation and activated Prismic token
 
 ## [1.0.5] - 2010-06-01
+
 - Removed footer border
 
 ## [1.0.4] - 2010-06-01
+
 - Fixed meter matching
 
 ## [1.0.3] - 2010-06-01
+
 - Fixed header on tablet and update social image
 
 ## [1.0.2] - 2010-06-01
+
 - Added sitemap
 
 ## [1.0.1] - 2010-06-01
+
 - Added og:image
 
 ## [1.0.0] - 2010-06-01
+
 - Rebuilt the project with Nuxt.js and Prismic.io 🎉